+++ src/utils/isEmail/isEmail.test.ts
@@ -1,0 +1,16 @@
+import isEmail from './isEmail';
+
+describe('isEmail', () => {
+  it('should be able to validate an email', () => {
+    expect(isEmail('test@email.com')).toBeTruthy();
+  });
+  it('should be able to validate an invalid email', () => {
+    expect(isEmail('testemail.com')).toBeFalsy();
+  });
+  it('should be able to validate an email with spaces', () => {
+    expect(isEmail('teste @ email. com')).toBeFalsy();
+  });
+  it('should be able to validate an empty email', () => {
+    expect(isEmail('')).toBeFalsy();
+  });
+});